integrantes

+++ ParcialPoo/ParcialPoo/Entidades/Ventas.cs
@@ -8,7 +8,7 @@ namespace ParcialPoo.Entidades
 {
     class Ventas
     {
-        private double identificador_producto;
+        private int identificador_producto;
         private string nombre_producto;
         private string descripcion;
         private double precio;
@@ -18,7 +18,7 @@ namespace ParcialPoo.Entidades
         {
         }
 
-        public Ventas(double identificador_producto, string nombre_producto, string descripcion, double precio, int cantidad)
+        public Ventas(int identificador_producto, string nombre_producto, string descripcion, double precio, int cantidad)
         {
             this.Identificador_producto = identificador_producto;
             this.Nombre_producto = nombre_producto;
@@ -27,7 +27,7 @@ namespace ParcialPoo.Entidades
             this.Cantidad = cantidad;
         }
 
-        public double Identificador_producto { get => identificador_producto; set => identificador_producto = value; }
+        public int Identificador_producto { get => identificador_producto; set => identificador_producto = value; }
         public string Nombre_producto { get => nombre_producto; set => nombre_producto = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
         public double Precio { get => precio; set => precio = value; }
